backend/requirements.txt
fastapi
uvicorn
python-multipart
openai==0.28
python-dotenv
aiofiles